♻️ use existing constants
needed to move things around to avoid dependency order issue with sourcemaps
