+++ packages/core/src/domain/transportConfiguration.ts
@@ -1,4 +1,4 @@
-import { BuildEnv, BuildMode, Datacenter, INTAKE_SITE } from '../boot/init'
+import { BuildEnv, BuildMode } from '../boot/init'
 import { includes } from '../tools/utils'
 import { TransportConfiguration, UserConfiguration } from './configuration'
 
@@ -18,7 +18,19 @@ const ENDPOINTS = {
   },
 }
 
-const CLASSIC_ALLOWED_SITES = ['datadoghq.com', 'datadoghq.eu']
+export const Datacenter = {
+  EU: 'eu',
+  US: 'us',
+} as const
+
+export type Datacenter = typeof Datacenter[keyof typeof Datacenter]
+
+export const INTAKE_SITE = {
+  [Datacenter.EU]: 'datadoghq.eu',
+  [Datacenter.US]: 'datadoghq.com',
+}
+
+const CLASSIC_ALLOWED_SITES = [INTAKE_SITE[Datacenter.US], INTAKE_SITE[Datacenter.EU]]
 
 type IntakeType = keyof typeof ENDPOINTS
 type EndpointType = keyof typeof ENDPOINTS[IntakeType]
