+++ packages/core/src/index.ts
@@ -4,7 +4,6 @@ export { computeStackTrace } from './domain/tracekit'
 export {
   BuildEnv,
   BuildMode,
-  Datacenter,
   defineGlobal,
   makePublicApi,
   commonInit,
@@ -46,3 +45,4 @@ export { createContextManager } from './tools/contextManager'
 export { limitModification } from './tools/limitModification'
 
 export * from './tools/specHelper'
+export { Datacenter } from './domain/transportConfiguration'
